+++ tests/test_deployment.py
@@ -2,18 +2,8 @@ import dataclasses
 
 import numpy as np
 import pytest
-from distributed.utils_test import client, loop, cluster_fixture, loop_in_thread, cleanup # noqa: F401
-
 
 import znflow
-
-@pytest.fixture
-def vanilla_deployment():
-    return znflow.deployment.VanillaDeployment()
-
-@pytest.fixture
-def dask_deployment(client):
-    return znflow.deployment.DaskDeployment(client=client)
 
 @znflow.nodify
 def compute_sum(*args):
